Move find_by_selector_in_regions to view
